refactor: 迁移结构域注册
